# `F.avg_pool3` CUDA backward: gpuAtomicAddNoReturn -> fastAtomicAdd